services/job-scraper-service/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
aiohttp>=3.9.1

# AI/ML
google-generativeai>=0.3.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2  # For FastAPI TestClient
pytest-cov>=4.1.0

# Logging and monitoring
structlog>=23.2.0

# Database
motor>=3.3.2  # Async MongoDB driver
pymongo>=4.6.1
supabase>=2.3.1  # Supabase client

# Common utilities
python-dateutil>=2.8.2
tenacity>=8.2.3  # For retries
python-json-logger>=2.0.7  # For structured logging
